requirements.txt
torch==1.12.1
torchvision==0.13.1

lightning==2.1.4
pytorch-lightning==2.1.0  # Included because it was in the full list and you imported pytorch_lightning.loggers
monai==1.3.2
torchio==0.19.7
wandb==0.17.7
easydict==1.13
tqdm==4.66.4

numpy==1.26.4  
scipy==1.13.1
pandas==2.2.2
tifffile==2024.6.18
scikit-image==0.24.0

opencv-python==4.10.0.84
matplotlib==3.9.0
Pillow==10.3.0  
PyYAML==6.0  